Create README.md
STM8S103F LCD clock with alarm settings for "daily"' , "day of the week" and "date of the year" options. The buzzer is connected to the PD4 and beeps for approximately 1 minute on alarm set time matching clock time. This can be done on daily basis if "daily" is selected in alarm menu. The same can be done for repetitive alarm for "day of week" and on date basis also. For entering alarm menu press and hold "NEXT" button first then while holding "NEXT" button press and release "MENU" button. After "ALARM MENU" is displayed on screen proceed with adjusting alarm time with "NEXT" and "SAVE" buttons. Pressing MENU button only ,system enters clock time adjustment menu which can be done with "NEXT" and "SAVE" buttons.